BACKLOG-23333: Move the .tgz to dist/ folder (#154)
Update the CI to get the .tgz file from the dist/ folder.
In case there is no .tgz file in the dist/ folder, fallback to the old behaviour to smoothly release this change. Eventually the else condition should be removed once all Javascript projects use the "new" build tooling (that generates the file in the dist/ folder).