ci(npm): Update bundling of NPM modules (#203)
Update the the building of NPM modules to use dist/package.tgz for the file path of the generated tgz file.
If not found, use the previous location (dist/<<project>>-v<<version>>.tgz).
Also remove the "Package the module" step as it duplicates the "Build the module" one, that already performs a yarn build.